pagedgeometry和ogitor一致版本

+++ kbe/src/client/pagedgeometry/include/RandomTable.h
@@ -91,7 +91,7 @@ inline void RandomTable::generateRandomNumbers()
 	MTRand mtrand(time(NULL));
 #endif
 	for(unsigned long i = 0; i < tableSize; i++)
-		table[i] = (float)mtrand.rand();
+		table[i] = mtrand.rand();
 #endif //USE_OGRE_RANDOM
 }
